AS-IS) WebSecurityConfig 권한 수정, 파라미터 required false 설정 완료 TO-BE) redis token 개발 시작

+++ src/main/java/com/spring/sharepod/service/BoardService.java
@@ -36,7 +36,7 @@ public class BoardService {
 
     //메인 페이지 전체 게시글 불러오기
     @Transactional
-    public List<BoardAllResponseDto> getAllBoard(int limitcount) {
+    public List<BoardAllResponseDto> getAllBoard(Long limitcount) {
 
         // 모든 게시글 가져오기
         List<Board> boardList = boardRepository.findAllByOrderByCreatedAtDesc(limitcount);
